- Modification du temps de simulation (double -> long) - Ajout de safe checks dans les setters (defensive programming) - Mise à jour des diagrammes uml - Correction d'un problème avec la configuration

+++ src/main/java/be/simulation/evenements/FinDeSimulation.java
@@ -13,7 +13,7 @@ public class FinDeSimulation extends Evenement {
 	/**
 	 * Evenement de fin de simulation.
 	 */
-	public FinDeSimulation(final Double tempsPrevu) {
+	public FinDeSimulation(final long tempsPrevu) {
 		super(tempsPrevu);
 	}
